requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers==0.33.1
transformers==4.46.3
accelerate==1.6.0
xformers>=0.0.22
spconv-cu118>=2.3.0

# UI and visualization
gradio==5.25.2
matplotlib>=3.7.0

# Image processing
numpy==1.26.4
opencv-python
pillow>=10.0.0
einops
scikit-image
kornia==0.8.0
timm==0.6.7
sentencepiece
ftfy

# 3D processing
trimesh>=4.0.0

# Utilities
huggingface_hub>=0.19.0
python-dotenv>=1.0.0
